Allow compiling the Linux module on case-insensitive filesystems
When compiling the kernel modules, .o files with different cases are
generated. Unfortunately, that does not work on case-insensitive
filesystems, e.g. when VBoxPci.o wants to be linked into vboxpci.o but
is deleted before that can happen because ld removes vboxpci.o before
writing that file (and vboxpci.o *is* VBoxPci.o on such filesystems).

This problem occurs for VBoxNetAdp.o, VBoxNetFlt.o and VBoxPci.o.

As we are about to add a Vagrantfile which enables us to run Linux on
MacOSX and Windows (but uses the underlying filesystem), we need to
address this problem first.

So let's just give the offending .c files a -common.c suffix instead.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>